Application ready. Edition 2.4
