--- src/main/java/com/example/cookapp/service/RecipeInterface.java
@@ -1,18 +1,0 @@
-package com.example.cookapp.service;
-
-import com.example.cookapp.model.Recipe;
-
-import java.util.Collection;
-
-public interface RecipeInterface {
-
-    Recipe addRecipe(Recipe recipe);
-
-    Recipe getTheRecipe(Integer id);
-
-    boolean deleteTheRecipe(Integer id);
-
-    Recipe editTheRecipe(Integer id, Recipe recipe);
-
-    Collection<Recipe> getAllRecipe();
-}
